# *: improve 'insert into select' performance by optimize utf8 charset check