--- a/table/column.go
+++ b/table/column.go
@@ -315,21 +315,23 @@ func CastValue(ctx sessionctx.Context, val types.Datum, col *model.ColumnInfo, r
 	str := casted.GetString()
 	utf8Charset := col.Charset == mysql.UTF8Charset
 	doMB4CharCheck := utf8Charset && config.GetGlobalConfig().CheckMb4ValueInUTF8
-	for i, w := 0, 0; i < len(str); i += w {
-		runeValue, width := utf8.DecodeRuneInString(str[i:])
-		if runeValue == utf8.RuneError {
-			if strings.HasPrefix(str[i:], string(utf8.RuneError)) {
-				w = width
-				continue
+	if !utf8.ValidString(str) {
+		for i, w := 0, 0; i < len(str); i += w {
+			runeValue, width := utf8.DecodeRuneInString(str[i:])
+			if runeValue == utf8.RuneError {
+				if strings.HasPrefix(str[i:], string(utf8.RuneError)) {
+					w = width
+					continue
+				}
+				casted, err = handleWrongCharsetValue(ctx, col, &casted, str, i)
+				break
+			} else if width > 3 && doMB4CharCheck {
+				// Handle non-BMP characters.
+				casted, err = handleWrongCharsetValue(ctx, col, &casted, str, i)
+				break
 			}
-			casted, err = handleWrongCharsetValue(ctx, col, &casted, str, i)
-			break
-		} else if width > 3 && doMB4CharCheck {
-			// Handle non-BMP characters.
-			casted, err = handleWrongCharsetValue(ctx, col, &casted, str, i)
-			break
+			w = width
 		}
-		w = width
 	}
 
 	if forceIgnoreTruncate {