<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

Insert into select is very slow ...

```
CREATE TABLE `xx` (
  `id` int(11) DEFAULT NULL,
  `c` varchar(512) DEFAULT NULL
)
```
```
CREATE GLOBAL TEMPORARY TABLE `t` (
  `id` int(11) DEFAULT NULL,
  `c` varchar(512) DEFAULT NULL
)
```
prepare some data:

```
insert into xx values (1, repeat(‘x’, 512));
insert into xx select * from xx;  // execute several times
insert into xx select * from xx limit yy;
```

```
explain analyze insert into t select * from xx;
time:5.06s, prepare:3.1s, insert:1.95s  552.2 MB  1062144
```

As you can see, the total time is 5.06s, the actual insert cost **1.95s**, which prepare the data takes **3.1s** ...

![image](https://user-images.githubusercontent.com/1420062/126446777-e6935cb7-3c45-480d-a448-a84cfdeb4542.png)

A lot of the time is spend on the `CastValue` operation.

Issue Number: close #xxx <!-- REMOVE this line if no issue to close -->

Problem Summary:

### What is changed and how it works?


What's Changed:


In `CastValue`, we check the datum to be utf8mb4 encoding, it takes a lot of time.

I prefer to do that using the standard lib `utf8.ValidString(str)` for fast checking.
If it return false (which should be rare), we use the old way to check.

How it Works:

After this change, the time `insert into select` 550M down from **5s** to **3.32s**

And the flame graph now: 

![image](https://user-images.githubusercontent.com/1420062/126446104-741e29f0-dbcc-4600-a676-d189abcec95c.png)

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)

A benchmark test is added

```
cd session
 go test -bench BenchmarkInsertIntoSelect -run XXX -benchmem -cpuprofile cpu.out -benchtime 30s
```

Before vs After:

```
BenchmarkInsertIntoSelect-16    	     319	 112433885 ns/op	103542845 B/op	  365614 allocs/op
BenchmarkInsertIntoSelect-16    	     486	  73791669 ns/op	103508584 B/op	  365436 allocs/op
```

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note <!-- bugfixes or new feature need a release note -->

- No release note