@tiancaiamao 2021-08-03
Close #25745

A regression test in the test case [show](http://www.zenlife.tk:18081/) the changes:
![canvas](https://user-images.githubusercontent.com/1420062/127992955-13fa9948-4752-468e-8c84-9031c1dd8dfa.png)